Implement basic support for Unix FD passing
This introduces two new types, UnixFD and UnixFDIndex, of which only the
latter is actually supported by the wire protocol. However, all methods
of Connection automatically convert between these types, provided that the
transport supports it.

This also adds the new internal transport interface, which handles the
transport-specific operations of a connection.