Handle method calls that lack an interface
The specification states that if the interface header is missing for a
method call, but the method name is unique, the method must be called.
If the method name is not unique, the behaviour is undefined. We take
the easy way out and always call the method on the first interface that
is found when iterating over the map.