Handle calls on closed connections better
Previously, calls on a closed connections would panic, mostly because of
sends on the closed output channel. Now, a special error is returned if
a connection is already closed. This requires a RWMutex to protect the
channel from being closed while a method tries to send to it. It also
changes some methods to returns errors.