Seperate private and shared connections
This mirrors the API of libdbus. Shared connections as returned by
SessionBus and SystemBus are already registered and authenticated,
and should be used in most cases. For private connections obtained with
Dial or NewConn, authentication can be controlled. As a result, Auth and
Hello are exported, and Auth can be passed a list of auth mechanisms to
try. The interface for authentication has changed as well.

Also, the new functions {Session,System}BusPrivate return private
connections to the respective message bus.

Updates #18.