auth: Send uid rather than username when authenticating
There is no particular advantage to sending the username. It just
means the other end has to map it to a uid when comparing with the
uid from getpeercred. Additionally that mapping is not always
possible, as the remote side may not have the correct /etc/passwd.
For instance, this can happen when talking to a dbus service inside
a container.