Fix encoding alignment logic.
The newEncoder method is recursive, but alignment logic must be based on
"global" absolute position. This FIX allows recursive calls to set the
initial position.

Unit tests added to very the problem and fix.

The problem was discovered for nested containers where the start of the
nested container can be on a 4-byte-aligned (but not 8-byte-aligned) boundary.