conn: fix RemoveSignal and add test
RemoveSignal had previously used `range conn.signals`. This
is not safe as the array is modified during the loop. The
modified RemoveSignal using a decrementing loop is curtousy
of icza: http://stackoverflow.com/a/29006008

The test creates a *Conn with nil reader and adds three copies
each of two signal channels. It then removes the first and
checks that only the second chan is present.