Refactor export code to allow a method table to be provided
Allow a table of function closures to be substituted for a native
object during export. This allows exporting generated code onto
dbus more easily. This is especially handy when interfacing with
some generic service that wishes to expose dbus methods based on
some definition language.

This refactor also cleans up the logic for the override table a bit.
The exported objects are now stored as a table of method name to reflect.Value
where the reflect.Value has been filtered to only be valid dbus methods.
This is done at export time instead of doing the lookup dynamically at
call time.