Refactor connection implementation details
This refactor allows for all the details of connection message
handling to be overridden. The previously inline handlers for signals
and messages are now implemented as an instance of the interfaces.

This is useful for substituting the default object model for another
model allowing full customization of object and method lookup. This
allows for higher level abstractions to be built without breaking
compatibility with the current library.

All '*Conn' Export and Signal registration will only work with these
default handlers. It is expected that implementations wishing to
override the behavior of method lookup and signal delivery will
implement similar mechanisms. The details of what can be exported or
how a signal should be delivered are up to the implementation of the
interface and thus should not be part of the interface definition.