dbus: Only dereference dest in store if Kind is Interface or Ptr.
Trying to store the result of a call with a signature like like
`a{sa{sv}}` to a `map[string]map[string]Variant` panics due to trying to
call Elem() on a Map Value.

Only call Elem() when the destination value is of Kind Interface or Ptr.

An example of such a call:
  https://developer.gnome.org/NetworkManager/0.9/spec.html#org.freedesktop.NetworkManager.Settings.Connection.GetSettings