Change the DBusError interface
Since The DBusError interface requires one to return a concrete type
from the D-Bus library, that library must be imported by every thing
that returns one of these errors. It would be friendlier to support
returning base types so that the implementations need not import the
godbus library.

This is not compatible, but the interface is still fairly new.