Add support for more of go's basic types
This significantly widens the API for Store. The change is backwards
compatible, and now allows storeing into dest values to which the
source is convertible. This includes things like a type Boolean which
is has a bool as an underlying type, but also anything that can be
converted in go, i.e. various numeric types.

Slight modifications were required to the encoder and decoder to allow
for int and uint to be used as basic types on D-Bus.

This now provides the full power of the go type system to be used over
D-Bus transparently.

Fixes #53