Allow implementations to decide when to make signal handling asynchronous.
This moves the place where the signal handler is asynchronous. For signals
the ordering of arrivial may be important. Making them asynchronous by defualt
does not allow for this policy decision to be made by the signal handler. In
order to preserve the current behavior, make the default handler asynchronous
instead of all handlers.