Fix decoder for arrays of structs
Arrays of structs are a special case for the decoder. Since it needs
to decode structs as an array of interface{} values instead of a known
type first, but we also want to allow decoding of an array of
interface{} values and the two have different alignment offsets we
need to account for the difference in the array decoding.

Fixes #85