Update server example
The server example would throwaway the first message received on the
message bus if the server.go file were to be dbus activated.

By changing the order of the code it would receive and execute the
correct method provided by the interface instead of throwing it away
when noting mapped to it.

Signed-off-by: mattemagikern <mansgariusson@gmail.com>