Allow the Store for Variants and Properties.
This adds a Store method to Variants so that they may be easily
unmarshalled into native go types. It also adds a StoreProperty method
that can be used to directly unmarshal a value returned from
org.freedesktop.DBus.Properties.Get instead of requiring it go through
a Variant first.

Fixes #197