Switch execution to x/sys/execabs for windows security fix
This may or may not be relevant. I don't know how much it makes sense (or is possible) to use dbus on Windows, but I wanted to be nice and open this as part of the efforts in https://github.com/fyne-io/fyne/pull/2344.

The os/exec package on Windows will match the behaviour of cmd.exe by considering the local folder as a primary part of the path. This means that a malicious binary with the same name, in the current folder, would be run instead of the expected binary in the system path. Due to the backwards compat being an issue, this could not be fixed within ox/exec before Go v2. See https://blog.golang.org/path-security for more info.