do not send data with DBUS auth mechanisms
When running inside a container using systemd dynamic user to run as an
actual non-root user outside the container (i.e. UID 12345), the DBUS
refuses to allow the godbus code to connect. This is because the godbus
code sends the UID of the user.

The sdbus C library sends "AUTH EXTERNAL\r\n", whereas godbus sends
"AUTH EXTERNAL 30\r\n". Since godbus goes the extra step to tell DBUS
that the current UID is 0 (hex 0x30, getuid() == 0, i,e, root), DBUS
refuses to allow the process to connect because it sees the process is
actually running as UID 12345 (for example) outside of the container.

When the UID is not specified by just sending "AUTH EXTERNAL\r\n", DBUS
sends back a line of "DATA\r\n" and expects a reply of "DATA\r\n". The
godbus code specifically checks for "DATA" when waitingForOk in
tryAuth(), but sends a "CANCEL\r\n" to the DBUS and errors out.

Instead of erroring out, the tryAuth() code in godbus should reply with
"DATA\r\n" when the status is 'waitingForOk' and the line read back from
the DBUS begins with "DATA".

Fixes #264