prop: add the Prop.Introspection() method
Sometimes it is convenient to generate the introspection data for properties
before they are exported. For example, when the introspection data for an
object is gathered before the corresponding interfaces are exported.

Move the relevant pieces from Properties.Introspection() into the newly
created Prop.Introspection() method, which generates the introspection
data for the corresponding Prop instance using the provided property
name.