Introduce and use {Une,E}scapeBusAddressValue
D-Bus specification [1] requires that the values in server address need
to be escaped in a special way, and other clients perform the needed
escaping (e.g. systemd [2] does that).

More to say, if dbus sees a character that should have been escaped, it
returns an error [3].

Add EscapeBusAddressValue function (with some tests and a benchmark),
and use it from tryDiscoverDbusSessionBusAddress().

Add UnescapeBusAddressValue, and use it from getKey().

The functions are exported as they might be of use to other packages
(in particular, runc [4]).

NOTE that the Escape... function does not escape '*' symbol, because
de-facto it is the list of optionally-escaped characters since 2007, but
that is not yet documented in D-Bus spec (see [5] for more details).

[1] https://dbus.freedesktop.org/doc/dbus-specification.html#addresses
[2] https://github.com/systemd/systemd/blob/5efbd0bf897a990ebe43d7dc69141d87c404ac9a/src/libsystemd/sd-bus/bus-internal.c#L294-L318
[3] https://gitlab.freedesktop.org/dbus/dbus/-/blob/37b76d13738e782fe2eb12abdd0179745c0b3f81/dbus/dbus-address.c#L330
[4] https://github.com/opencontainers/runc/pull/3356
[5] https://gitlab.freedesktop.org/dbus/dbus/-/merge_requests/248

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>