homedir: revamp
Back in the day os/user package was relied on libc functions to parse
/etc/passwd and /etc/group. This was fixed in Go 1.11 by [1] and
subsequent fixes, so there is no more need to parse /etc/passwd.

This assumes that users know when and how to use osusergo build tag
(since this is not something very new, most are).

It might make sense to note in the changelog / release notes that the
static_build tag is no longer supported, and osusergo tag should be
used instead.

[1] https://go-review.googlesource.com/c/go/+/92456

Updates: da6051ac2b
Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>