Reduce bytes to string allocs in decoder
Instead of allocating when converting
small byte slices to strings in the decoder,
we can write to a 4kb buffer and use unsafe.String
which refers to that buffer.
This approach was successfully tested in
https://github.com/marselester/systemd which is used
by https://github.com/parca-dev/parca-agent.

Note, BenchmarkUnixFDs showed 8 allocs reduction.
With larger dbus messages, the savings should be significant.