Atomic function to get an exported object for a path
GetOrAddExportedObject returns either an existing object or a newly
created object for a specific path. This will solve the case where
we could get a mixup when two different goroutines competed to
create an object for a specific path. The result might still be
unsatifying but atleast its either or and not a mix.