Add ConnOption WithDefaultSendTimeout
Added a default timeout for a connection, as well as a ConnOption to
set this value. If a default timeout has been set, all calls from that
connection will be aborted if the timeout is exceeded unless we have
specified the context.

If we always want to have the same timeout for all calls from a connection,
it would be easier if the timeout could be set once to the
connection itself rather than always having to use CallWithContext.

Change-Id: I49f34caeeace5b2cf29e56ef97f7a4817d15d158