Call and Go use context.TODO() instead of context.Background()
Instead of using a bool to determine if we have specified the context,
we now assume that context.TODO() is only sent from functions where the
user didn't specify a context. If we have context.TODO() and a default
timeout we will use that to create our context and cancelFunc, otherwise
we will use context.Background() or the provided context.

Change-Id: I043aae124a92ba55beac0f380dc8ad3a44bbf382