ci: add stable and oldstabe go to the matrix, drop old ones
Use "oldstable" (currently 1.25) and "stable" (currently 1.26) aliases,
so we always test the supported Go versions.

Drop the 1.21, 1.22, and 1.23 from the matrix as they are no longer
maintained / supported.

Finally, add a hack to test against the Go version specified in go.mod,
so we always test against a minimally supported version. For this to
work, we had to checkout sources before installing Go.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>