Fix and simplify getHomeDir
Since commits 99fac80 and ec919d8 it's clear that getHomeDir is only
used on Windows. Ironically, $HOME is not used on Windows (instead,
it's %USERPROFILE%), so we always fallback to user.Current, which
internally uses os.UserHomeDir to fill in HomeDir.

So, let's just use os.UserHomeDir directly (and fall back to "/" in case
of an error like the previous implementation did).

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>