Revert "Use pointer semantics in depthCounter"
This reverts commit 59425418c51ba6584c944b7659b1ab5c59c1f856.

Pointer semantics on the caller are actually incorrect here: the
returned depthCounter is passed on to a deeper validation function, but
that shouldn't modify the original depthCounter from the caller. With
pointer semantics, we don't count e.g. the depth of array nesting, but
just how many arrays are present in the signature. For structs this is
especially bad since EnterStruct is called for each struct field, so
this unintentionally limited the number of struct fields that could be
present in a signature to 32.

Fixes #425.